pyproject.toml
[tool]

[tool.poetry]
name = "zillionare-pyqmt"
version = "0.1.0"
homepage = "https://github.com/zillionare/pyqmt"
description = "Skeleton project created by Python Project Wizard (ppw)."
authors = ["aaron yang <aaron_yang@jieyu.ai>"]
readme = "README.md"
license =  "MIT"
classifiers=[
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
]

packages = [
    { include = "pyqmt" },
    { include = "tests", format = "sdist" },
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
fire = "0.4.0"

black  = { version = "^22.3.0", optional = true}
isort  = { version = "5.10.1", optional = true}
flake8  = { version = "4.0.1", optional = true}
flake8-docstrings = { version = "^1.6.0", optional = true }
pytest  = { version = "^7.0.1", optional = true}
pytest-cov  = { version = "^3.0.0", optional = true}
tox  = { version = "^3.24.5", optional = true}
mkdocs  = { version = "^1.2.3", optional = true}
mkdocs-include-markdown-plugin  = { version = "^3.2.3", optional = true}
mkdocs-material  = { version = "^8.1.11", optional = true}
mkdocstrings  = { version = "^0.18.0", optional = true}
mkdocs-material-extensions  = { version = "^1.0.3", optional = true}
twine  = { version = "^3.8.0", optional = true}
mkdocs-autorefs = {version = "^0.3.1", optional = true}
pre-commit = {version = "^2.17.0", optional = true}
toml = {version = "^0.10.2", optional = true}
livereload = {version = "^2.6.3", optional = true}
pyreadline = {version = "^2.1", optional = true}
mike = { version="^1.1.2", optional=true}
mypy = {version = "^1.5.1", optional = true}
setuptools = {version="^68.0", optional = true}
pkginfo = {version="^1.9", optional = true}
virtualenv = {version="^20.0", optional = true}
charset-normalizer = "^3.1.0"  # Needed cause of a bug in poetry
urllib3 = "^1.26.16"  # Needed cause of a bug in poetry
blacksheep = "^2.0.0"
uvicorn = "^0.24.0.post1"
numpy = "^1.26.2"
pandas = "^2.1.3"
apscheduler = "^3.10.4"
cfg4py = "^0.9.4"
arrow = "^1.3.0"
clickhouse-connect = "^0.6.22"
zillionare-core-types = "^0.6.3"
sqlite-web = "^0.6.1"
redis = "^5.0.1"
hiredis = "^2.3.2"
[tool.poetry.extras]
test = [
    "pytest",
    "pytest-cov"
    ]

dev = ["tox",
    "pre-commit",
    "twine",
    "toml",
    "black",
    "isort",
    "flake8",
    "flake8-docstrings",
    "mypy"
    ]

doc = [
    "mkdocs",
    "mkdocs-include-markdown-plugin",
    "mkdocs-material",
    "mkdocstrings",
    "mkdocs-material-extension",
    "mkdocs-autorefs",
    "mike",
    "setuptools",
    "pkginfo",
    "virtualenv"
    ]

[tool.poetry.scripts]
pyqmt = 'pyqmt.cli:main'

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "primary"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | \.history
)/
'''
[tool.isort]
profile = "black"
